% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjRank.R
\name{catvar2binvars}
\alias{catvar2binvars}
\title{Encode categorical variable to series of binary indicators}
\usage{
catvar2binvars(catvar, prefix)
}
\arguments{
\item{catvar}{Categorical variable vector}

\item{prefix}{name of variable / prefix to be used for naming of output}
}
\value{
Matrix of binary indicators, one column per category
}
\description{
Encode categorical variable to series of binary indicators
}
\keyword{internal}
