% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjRank.R
\name{get_order}
\alias{get_order}
\title{Function or compute feature ordering}
\usage{
get_order(data.bin, data.var, outcome, taskid, attr.vars, min_obs)
}
\arguments{
\item{data.bin}{Matrix of binary features}

\item{data.var}{Matrix showing variance by attribute within tasks}

\item{outcome}{Outcome vector}

\item{taskid}{Task ID vector}

\item{attr.vars}{Character vector of categorical attribute names}

\item{min_obs}{Minimum number of observations needed for ranking.}
}
\value{
Ordering
}
\description{
Function or compute feature ordering
}
\keyword{internal}
