% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjRank.R
\name{taskvariance}
\alias{taskvariance}
\title{Compute binary variance indicator}
\usage{
taskvariance(catvar, prefix, taskid)
}
\arguments{
\item{catvar}{Categorical variable vector}

\item{prefix}{name of variable / prefix to be used for naming of output}

\item{taskid}{Categorical task ID vector}
}
\value{
A vector indicating variance of attribute in a task (0,1)
}
\description{
Compute binary variance indicator
}
\keyword{internal}
