% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjRank.R
\docType{data}
\name{ukraine}
\alias{ukraine}
\title{Example conjoint data in cjRank}
\description{
A subset of the data used in Dill, Janina, Howlett, Marnie, & 
Muller-Crepon, Carl (2022). At Any Cost: How Ukrainians Think about 
Self-Defense Against Russia. _American Journal of Political Science_, forthcoming.
}
\keyword{data}
